mcps/super-claude/pyproject.toml
[project]
name = "super-claude"
version = "0.1.0"
description = "Super Claude MCP - Auth, Filesystem, Shell, Docker tools"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "uvicorn>=0.30.0",
    "onepassword-sdk>=0.1.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "supernotelib>=0.6.0",
    "reportlab>=4.0.0",
    "lxml>=5.0.0",
    "pymupdf>=1.24.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["plugins"]
